Move estimateUnlockAndSwap function
